# docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "quickstart-restful-user-profile-crud",
        "description": "Minimal Spring Boot based RESTful CRUD example, which manages user profile entity (including Swagger UI).",
        "version": "23.0.0-SNAPSHOT"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "tags": [
        {
            "name": "User profiles",
            "description": "API endpoints for managing user profile entity."
        }
    ],
    "paths": {
        "/api/v1/user-profiles/{id}": {
            "get": {
                "tags": [
                    "User profiles"
                ],
                "description": "Retrieve user profile by id.",
                "operationId": "findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidContentTypeMeta"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidHttpMethodMeta"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User profiles"
                ],
                "description": "Updates an existing user profile.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserProfileDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidContentTypeMeta"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidHttpMethodMeta"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User profiles"
                ],
                "description": "Delete user profile.",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidContentTypeMeta"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidHttpMethodMeta"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "enum": [
                                        "100 CONTINUE",
                                        "101 SWITCHING_PROTOCOLS",
                                        "102 PROCESSING",
                                        "103 EARLY_HINTS",
                                        "103 CHECKPOINT",
                                        "200 OK",
                                        "201 CREATED",
                                        "202 ACCEPTED",
                                        "203 NON_AUTHORITATIVE_INFORMATION",
                                        "204 NO_CONTENT",
                                        "205 RESET_CONTENT",
                                        "206 PARTIAL_CONTENT",
                                        "207 MULTI_STATUS",
                                        "208 ALREADY_REPORTED",
                                        "226 IM_USED",
                                        "300 MULTIPLE_CHOICES",
                                        "301 MOVED_PERMANENTLY",
                                        "302 FOUND",
                                        "302 MOVED_TEMPORARILY",
                                        "303 SEE_OTHER",
                                        "304 NOT_MODIFIED",
                                        "305 USE_PROXY",
                                        "307 TEMPORARY_REDIRECT",
                                        "308 PERMANENT_REDIRECT",
                                        "400 BAD_REQUEST",
                                        "401 UNAUTHORIZED",
                                        "402 PAYMENT_REQUIRED",
                                        "403 FORBIDDEN",
                                        "404 NOT_FOUND",
                                        "405 METHOD_NOT_ALLOWED",
                                        "406 NOT_ACCEPTABLE",
                                        "407 PROXY_AUTHENTICATION_REQUIRED",
                                        "408 REQUEST_TIMEOUT",
                                        "409 CONFLICT",
                                        "410 GONE",
                                        "411 LENGTH_REQUIRED",
                                        "412 PRECONDITION_FAILED",
                                        "413 PAYLOAD_TOO_LARGE",
                                        "413 REQUEST_ENTITY_TOO_LARGE",
                                        "414 URI_TOO_LONG",
                                        "414 REQUEST_URI_TOO_LONG",
                                        "415 UNSUPPORTED_MEDIA_TYPE",
                                        "416 REQUESTED_RANGE_NOT_SATISFIABLE",
                                        "417 EXPECTATION_FAILED",
                                        "418 I_AM_A_TEAPOT",
                                        "419 INSUFFICIENT_SPACE_ON_RESOURCE",
                                        "420 METHOD_FAILURE",
                                        "421 DESTINATION_LOCKED",
                                        "422 UNPROCESSABLE_ENTITY",
                                        "423 LOCKED",
                                        "424 FAILED_DEPENDENCY",
                                        "425 TOO_EARLY",
                                        "426 UPGRADE_REQUIRED",
                                        "428 PRECONDITION_REQUIRED",
                                        "429 TOO_MANY_REQUESTS",
                                        "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
                                        "451 UNAVAILABLE_FOR_LEGAL_REASONS",
                                        "500 INTERNAL_SERVER_ERROR",
                                        "501 NOT_IMPLEMENTED",
                                        "502 BAD_GATEWAY",
                                        "503 SERVICE_UNAVAILABLE",
                                        "504 GATEWAY_TIMEOUT",
                                        "505 HTTP_VERSION_NOT_SUPPORTED",
                                        "506 VARIANT_ALSO_NEGOTIATES",
                                        "507 INSUFFICIENT_STORAGE",
                                        "508 LOOP_DETECTED",
                                        "509 BANDWIDTH_LIMIT_EXCEEDED",
                                        "510 NOT_EXTENDED",
                                        "511 NETWORK_AUTHENTICATION_REQUIRED"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user-profiles": {
            "get": {
                "tags": [
                    "User profiles"
                ],
                "description": "Get all the user profiles.",
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Page size",
                        "required": false,
                        "schema": {
                            "maximum": 200,
                            "minimum": 1,
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidContentTypeMeta"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidHttpMethodMeta"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageUserProfile"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User profiles"
                ],
                "description": "Create a new user profile.",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserProfileDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidContentTypeMeta"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseInvalidHttpMethodMeta"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Success",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfile"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorInvalidContentTypeMeta": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/InvalidContentTypeMeta"
                    }
                }
            },
            "ErrorResponseInvalidContentTypeMeta": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorInvalidContentTypeMeta"
                        }
                    }
                }
            },
            "InvalidContentTypeMeta": {
                "type": "object",
                "properties": {
                    "invalidType": {
                        "type": "string"
                    },
                    "supportedTypes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ErrorInvalidHttpMethodMeta": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/InvalidHttpMethodMeta"
                    }
                }
            },
            "ErrorResponseInvalidHttpMethodMeta": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorInvalidHttpMethodMeta"
                        }
                    }
                }
            },
            "InvalidHttpMethodMeta": {
                "type": "object",
                "properties": {
                    "invalidMethod": {
                        "type": "string"
                    },
                    "allowedMethods": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "ErrorObject": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "meta": {
                        "type": "object"
                    }
                }
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ErrorObject"
                        }
                    }
                }
            },
            "UserProfileDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
            "UserProfile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "email": {
                        "maxLength": 120,
                        "minLength": 5,
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
            "PageUserProfile": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserProfile"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}